backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
paramiko==3.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
websockets==14.1
loguru==0.7.3
